move filterPublicKeysFromUserInput to new account helper
